+++ components/src/lib/api/client.ts
@@ -1,3 +1,4 @@
+import type { SelectOptionType } from "flowbite-svelte/types";
 import { baseUrl } from "./config";
 
 export interface Candle {
@@ -98,4 +99,15 @@ export async function deleteNote(name: string): Promise<void> {
     const response: Response = await fetch(`${baseUrl}/api/notes/${name}`, {
         method: 'DELETE',
     });
+}
+
+export const resolutions: string[] = [ '1', '60', 'D', 'M' ];
+export const resolutionsOptions: SelectOptionType[] = [
+    ...resolutions.map((r: string): SelectOptionType => {
+        return { value: r, name: r };
+    })
+];
+export function symbolsToNoteName(symbols: string[]): string {
+    const sortedSymbols: string[] = symbols.sort();
+    return 'dashboard-' + sortedSymbols.join('-');
 }