+++ components/src/lib/components/chart.svelte
@@ -3,23 +3,22 @@
 	import { onMount, onDestroy } from 'svelte';
 
 	export let option: any;
+	export let height: string = '';
 
 	let chartDom: HTMLElement;
 	let chart: echarts.ECharts;
+	let loaded: boolean = false;
 
-	let innerWidth: number;
 	let innerHeight: number;
+	let innerWidth: number;
 
 	onMount(async function () {
-		chartDom.style.width = `${innerWidth - 20}px`;
-		chartDom.style.height = `${innerHeight - 20}px`;
+		if (height) chartDom.style.height = height;
+		if (!height) chartDom.style.height = `${innerHeight}px`;
 		chart = echarts.init(chartDom);
 		chart.showLoading();
-		window.onresize = function () {
-			chart.resize({
-				width: innerWidth,
-				height: innerHeight
-			});
+		window.onresize = () => {
+			chart.resize();
 		};
 	});
 
@@ -29,9 +28,10 @@
 
 	$: {
 		option;
-		if (chart && option) {
+		if (!loaded && chart && option) {
 			chart.hideLoading();
 			chart.setOption(option);
+			loaded = true;
 		}
 	}
 </script>